requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
attrs==25.3.0
blake3==1.0.5
cachetools==6.1.0
cbor2==5.6.5
certifi==2025.8.3
cffi==2.0.0b1
charset-normalizer==3.4.3
chz==0.3.0
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
cuda-bindings==13.0.0
cuda-pathfinder==1.1.0
cuda-python==13.0.0
cupy-cuda12x==13.5.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
einops==0.8.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.18.0
flashinfer-python==0.2.8
frozenlist==1.7.0
fsspec==2025.7.0
gguf==0.17.1
gpt-oss==0.1.0
h11==0.16.0
hf-xet==1.1.7
html2text==2025.4.15
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0rc0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kernels==0.9.0
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.4
mdurl==0.1.2
mistral_common==1.8.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
networkx==3.5
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.13.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.8.90
openai==1.99.6
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
outlines_core==0.2.10
packaging==25.0
partial-json-parser==0.2.1.1.post6
pillow==11.3.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.32.0rc2
psutil==7.0.0
py-cpuinfo==9.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.0a1
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.37.2
Pygments==2.19.2
pynvml==12.0.0
pytest==8.4.1
python-dotenv==1.1.1
python-json-logger==4.0.0.dev0
python-multipart==0.0.20
pytorch-triton==3.4.0+git11ec6354
PyYAML==6.0.2
pyzmq==27.0.1
ray==2.48.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.0
safetensors==0.6.2
scipy==1.16.1
sentencepiece==0.2.0
sentry-sdk==3.0.0a5
setproctitle==1.3.6
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==3.0.2
starlette==0.47.2
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.9.0.dev20250804+cu128
torchaudio==2.8.0.dev20250804+cu128
torchvision==0.24.0.dev20250804+cu128
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0+git663e04e8
triton_kernels==1.0.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uv==0.8.8
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.1+gptoss
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
xgrammar==0.1.21
yarl==1.20.1
zipp==3.23.0
